# lib/api-zod/src/generated/types/updateProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Order System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProductBody {
  name?: string;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  description?: string | null;
  unitPrice?: number;
  /** @nullable */
  stockQuantity?: number | null;
}